Add queries module and move query posts into there
