+++ api/src/routes/mod.rs
@@ -1,4 +1,5 @@
 pub mod catalogs;
 pub mod filters;
 pub mod objects;
+pub mod queries;
 pub mod users;
